+++ src/utils/helpers.ts
@@ -6,6 +6,8 @@ import { REF_COLLECTION, STATUS } from "~/constants/defaultValue";
 import subvn from "~/core/subvn";
 import POLICIES, { CORE_ACTION } from "~/modules/policy/policy.auth";
 import { PoliciesType, policyType } from "~/modules/policy/policy.modal";
+import { useAdapter } from "~/modules/auth/auth.hook";
+import { ADAPTER_KEY } from "~/modules/auth/constants";
 
 export const getPaging = (response: any) => ({
   current: response.page,
@@ -339,3 +341,9 @@ export const permissionConvert = (query:any)=>(action: ActionPolicy, key: KeyPol
   } 
   return POLICIES[ objj[ get(query,'refCollection') as any ] ]as [string,policyType] 
 };
+export const useIsAdapterSystem = () => {
+  const adapter = useAdapter();
+  const isAdapterSystem = useMemo(() => adapter === ADAPTER_KEY.STAFF, [adapter]);
+  return !!isAdapterSystem; // return true if adapter is system
+};
+