+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## [1.12.1](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/compare/v1.12.0...v1.12.1) (2024-04-29)
+
+
+### Bug Fixes
+
+* [#1467](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1467) action update, handle, permission, self test ([539d27d](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/539d27d9a70e30710978daba27a87f709bed4135))
+* [#1541](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1541) performance code ([305478e](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/305478ec4050b703d608d96d18ebbab51db65a06))
+* [#1541](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1541) performance package ([80e20c3](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/80e20c3e6d5071b3e74498145e8d8feda344afe3))
+* change field and show totalRevenue ([a4ca9a4](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/a4ca9a4a996eae04ec95b36b2c3342cd93713f0f))
+* performance lib ([9775e1c](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/9775e1c3b6c4d0582ed47a12907ea72fc7f4aff4))
+* show totalRevenue ([90c52a9](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/90c52a9a3067f57c617cb6ba94f70d3f295659bb))
+
+# [1.12.0](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/compare/v1.11.5...v1.12.0) (2024-04-29)
+
+
+### Bug Fixes
+
+* [#1522](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1522) CRUD, self test ([bd68183](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/bd681838aeea1378db83f1037022422ceb2f4aea))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) - add select employee seller ([64f7070](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/64f707052319045fb373502f8d1b186af2d3d45b))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) - change name file adress common ([c87ac20](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/c87ac2001401ff9fe9a0d50ed8c52fe7a37ffb11))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) - change name vi field urban type ([fa81514](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/fa8151426c06ba874bb57e3bf6970806400846ae))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) add api export error file ([f56f915](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/f56f915db39a47c60b2b1c09e608d81414bd7e3f))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) add infomation for pharmacy ([19ce91b](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/19ce91b236831f757e2c2b3bb1487d94b8571caf))
+* [#1533](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/issues/1533) form select employee seller ([0c71449](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/0c71449ec8cb67f8f15814d0c88eb1e41f4cc001))
+* fix file export from modal import pharmacy ([3be2001](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/3be2001ea210bd84527d688b623319fcbaa9352e))
+
+
+### Features
+
+* convert pharmacy ([aac0b5d](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/commit/aac0b5d3a045d613df96c11780c600ff63f02ac4))
+
 ## [1.11.5](https://gitlab.com/worldtech-ltd/worldpharma-dashboard/compare/v1.11.4...v1.11.5) (2024-04-12)